fix: 拆分还款管理与金融资讯页面路由
